Fix a bug with array freeing in hash table implementation
Some other things are in here which are distracting from the
main fix, which is line 174 of value.h. What was happening
was that findEntry was ending up in an infinite loop
when you ran the following Lox code:

```
var thing = 1;
```

The reason for this was that we were starting with an empty table,
and that required that we resize it, in our case up to 8:

https://github.com/chetaldrich/lox/blob/a1a827b067f1b76920faecbf52be79214a3f533b/clox/table.c#L43

This function was then running the FREE_ARRAY macro at the end:

https://github.com/chetaldrich/lox/blob/b546d6d75de1be57dd460b8f26645c3a4c5767bd/clox/memory.h#L15-L16

And that macro reset the values of everything. Turns out that I had
inadvertently set the default for the enum type Value to VAL_BOOL
instead of VAL_NIL, which was the cause of the bug.

The manifestation was that findEntry was then ending up in an infinite
loop because [this
check](https://github.com/chetaldrich/lox/blob/a1a827b067f1b76920faecbf52be79214a3f533b/clox/table.c#L28)
was never succeeding because the "empty" arrays in the table were
actually VAL_BOOL instead of VAL_NIL.

Now it works!